% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_dag.R
\name{condense_dag}
\alias{condense_dag}
\title{Condense a DAG into a rulegraph shape, but provide statistics on instance completion}
\usage{
condense_dag(dagfile, outfile = tempfile())
}
\arguments{
\item{dagfile}{path the dag file}

\item{outfile}{path to the output file to write a new .dot file into}
}
\description{
I had originally thought that I would end up with a rule graph with
numbers on the tail and head of each arrow saying how many jobs were done
and how many remained to be completed.  But I later decided to do dashed
nodes for completed, solid for incomplete.  Each such node will get a count
that says how many times, in total, that rule has been run, and the edges
between things will get a single weight that says how many such jobs
have been run, or are remaining to run.  So, there should never be an
edge from a solid (incomplete) parent to a dashed daughter, but we should see edges
from a dashed (complete) parent to a solid daughter.
}
\examples{
dagfile <- system.file("extdata/example_dag.dot", package = "SnakemakeDagR")
}
